Move the windows timezone detection to its own file we only import on Windows
